Add comments in code & export metrics.go from transport.go
