+++ gonvey.go
@@ -8,14 +8,12 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/Ullaakut/gonvey/logger"
-
 	"github.com/rs/zerolog"
 	"gopkg.in/tylerb/graceful.v1"
 )
 
 func main() {
-	log := logger.NewZeroLog(os.Stderr)
+	log := NewZeroLog(os.Stderr)
 	log.Info().Msg("gonvey is starting up")
 
 	config, err := GetConfig()
@@ -25,7 +23,7 @@ func main() {
 	}
 	config.Print(log)
 
-	zerolog.SetGlobalLevel(logger.ParseLevel(config.LogLevel))
+	zerolog.SetGlobalLevel(ParseLevel(config.LogLevel))
 
 	sig := make(chan os.Signal)
 	signal.Notify(sig, syscall.SIGINT, syscall.SIGTERM)
